requirements.txt
pyyaml
pytest
requests
jsonpath
pymysql
 allure-pytest